completed casper refactor
