+++ Prototype/Assets/Scripts/Items/Abstract Classes/Item.cs
@@ -5,15 +5,15 @@ public abstract class Item : MonoBehaviour
     protected int itemID;
     protected string itemName;
     protected string caption;
-    protected GameObject player;
-    protected PlayerStats stats;
+    protected GameObject casper;
+    protected Casper casperData;
 
     protected GameObject FlotingPointPrefab;
 
     protected virtual void Awake()
     {
-        player = GameObject.FindGameObjectWithTag("Player");
-        stats = player.GetComponent<PlayerStats>();
+        casper = GameObject.FindGameObjectWithTag("Player");
+        casperData = casper.GetComponent<Casper>();
         setDefaultInfo();
         gameObject.AddComponent<BoxCollider2D>();
         GetComponent<BoxCollider2D>().isTrigger = true;
