+++ Prototype/Assets/Scripts/Items/Active/AnxietyAttack.cs
@@ -12,7 +12,7 @@ public class AnxietyAttack : ActivatedItem
         float x = Random.Range(-1f, 1f);
         float y = Random.Range(-1f, 1f);
         Vector2 direction = new Vector2(x, y);
-        weapon = player.GetComponent<PointAndShoot>().getSelectedWeapon();
+        weapon = casper.GetComponent<PointAndShoot>().getSelectedWeapon();
         weapon.Shoot(direction);
     }
