+++ Prototype/Assets/Scripts/Player/Casper.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2201678267afe9a4e8584b8839295394
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
