+++ Prototype/Assets/Scripts/Player/Casper/Casper_Collision.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 15751cfdd43f2d6438824706c7835957
+guid: 4bf76fb88c5819c4681279a036615e93
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
