+++ Prototype/Assets/Scripts/Player/Casper/Casper_Health.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1aa21e206acd4254184ad726b28501f5
+guid: 8006c32992fea6345ac100e21ec1605a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
