+++ Prototype/Assets/Scripts/Player/Casper/Casper_Items.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9122b0e74d09a6640bf0a48bbfa4473e
+guid: 063ffebdcf5c8d440829508eb59f3675
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
