+++ Prototype/Assets/Scripts/Player/CasperData.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 32de8031b3ef8d744bf11d3b46d030cb
+guid: fc0a29694f6c7c946b32b8f72dfc4481
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
