+++ Prototype/Assets/Scripts/Player/PlayerData.cs
@@ -1,0 +1,8 @@
+﻿public class PlayerData
+{
+    public float totalShots = 0;
+    public int roomsCleared = -1;
+    public float bulletsHit = 0;
+    public int itemsPickedUp = 0;
+    public int enemiesKilled = 0;
+}
