+++ Prototype/Assets/Scripts/Player/PlayerData.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 105e9760af5511f49a86590a439608c8
+guid: f7361a40d8a315847bda2c35ea1affb2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
