+++ Prototype/Assets/Scripts/Utilities/SceneSingleton.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3b1a8e285b88548cdbfc1859bfb5791b
+guid: 701ff87019d5f7048a4bc90c68d78f4b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
